tests & refactoring

+++ backend/database/src/main/resources/db/migration/V0005__CreateQuestionnaireTable.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS Questionnaire (
     QuestionnaireID VARCHAR(50),
     CustomerID VARCHAR(50),
     Country VARCHAR(2),
-    MonthlyTurnover DOUBLE(32,2),
-    AnnualTurnover DOUBLE(32,2),
+    MonthlyTurnover NUMERIC(32,2),
+    AnnualTurnover NUMERIC(32,2),
     Reason VARCHAR(1000),
     Active BOOLEAN DEFAULT 1,
     FOREIGN KEY (CustomerID)
